Update threshold approach and rename files to not use `_jmh`
