+++ coloraide/spaces/hellwig.py
@@ -1,49 +1,64 @@
 """
-CAM02 class (JMh).
+Hellwig 2022: CAM16 class (JMh) with corrections.
 
+CAM16
 https://www.researchgate.net/publication/318152296_Comprehensive_color_solutions_CAM16_CAT16_and_CAM16-UCS
-https://en.wikipedia.org/wiki/CIECAM02
-https://www.researchgate.net/publication/221501922_The_CIECAM02_color_appearance_model
+https://www.researchgate.net/publication/220865484_Usage_guidelines_for_CIECAM97s
+https://doi.org/10.1002/col.22131
+https://observablehq.com/@jrus/cam16
 https://arxiv.org/abs/1802.06067
+
+CAM16 Corrections: Hellwig and Fairchild
+http://markfairchild.org/PDFs/PAP45.pdf
+
+Helmholtz Kohlrausch Effect extension: Hellwig, Stolitzka, and Fairchild
+https://www.scribd.com/document/788387893/Color-Research-Application-2022-Hellwig-Extending-CIECAM02-and-CAM16-for-the-Helmholtz-Kohlrausch-effect
 """
 from __future__ import annotations
 import math
+from .cam16 import (
+    M16,
+    MI6_INV,
+    M1,
+    adapt,
+    unadapt,
+    hue_quadrature,
+    inv_hue_quadrature
+)
+from .cam16 import Environment as _Environment
+from .lch import LCh
 from .. import util
 from .. import algebra as alg
-from ..spaces import Space, LChish
-from ..cat import WHITES, CAT02
+from ..cat import WHITES
 from ..channels import Channel, FLG_ANGLE
-from .lch import ACHROMATIC_THRESHOLD
 from ..types import Vector
-from .cam16_jmh import (
-    M1,
-    hue_quadrature,
-    inv_hue_quadrature,
-    eccentricity,
-    adapt,
-    unadapt
-)
-from .cam16_jmh import Environment as _Environment
 
-# CAT02
-M02 = CAT02.MATRIX
-M02_INV = [
-    [1.0961238208355142, -0.27886900021828726, 0.18274517938277304],
-    [0.45436904197535916, 0.4735331543074118, 0.07209780371722913],
-    [-0.009627608738429355, -0.00569803121611342, 1.0153256399545427]
-]
 
-XYZ_TO_HPE = [
-    [0.38971, 0.68898, -0.07868],
-    [-0.22981, 1.18340, 0.04641],
-    [0.00000, 0.00000, 1.00000],
-]
+def hue_angle_dependency(h: float) -> float:
+    """Calculate the hue angle dependency for CAM16."""
 
-HPE_TO_XYZ = [
-    [1.910196834052035, -1.1121238927878747, 0.20190795676749937],
-    [0.3709500882486886, 0.6290542573926132, -8.055142184361326e-06],
-    [0.0, 0.0, 1.0]
-]
+    return (
+        -0.160 * math.cos(h)
+        + 0.132 * math.cos(2 * h)
+        - 0.405 * math.sin(h)
+        + 0.080 * math.sin(2 * h)
+        + 0.792
+    )
+
+
+def eccentricity(h: float) -> float:
+    """Calculate eccentricity."""
+
+    # Eccentricity
+    h2 = 2 * h
+    h3 = 3 * h
+    h4 = 4 * h
+    return (
+        -0.0582 * math.cos(h) - 0.0258 * math.cos(h2)
+        - 0.1347 * math.cos(h3) + 0.0289 * math.cos(h4)
+        -0.1475 * math.sin(h) - 0.0308 * math.sin(h2)
+        + 0.0385 * math.sin(h3) + 0.0096 * math.sin(h4) + 1
+    )
 
 
 class Environment(_Environment):
@@ -82,16 +97,15 @@ class Environment(_Environment):
         """Calculate the adaptation of the reference point and related variables."""
 
         # Cone response for reference white
-        self.rgb_w = alg.matmul_x3(M02, xyz_w, dims=alg.D2_D1)
+        self.rgb_w = alg.matmul_x3(M16, xyz_w, dims=alg.D2_D1)
 
-        self.d_rgb = [(self.yw * (self.d / coord) + 1 - self.d) for coord in self.rgb_w]
+        self.d_rgb = [alg.lerp(1, self.yw / coord, self.d) for coord in self.rgb_w]
         self.d_rgb_inv = [1 / coord for coord in self.d_rgb]
-        self.rgb_cw = alg.multiply_x3(self.d_rgb, self.rgb_w, dims=alg.D1)
-        self.rgb_pw = alg.matmul_x3(alg.matmul_x3(XYZ_TO_HPE, M02_INV), self.rgb_cw)
 
         # Achromatic response
-        rgb_aw = adapt(self.rgb_pw, self.fl)
-        self.a_w = self.nbb * (2 * rgb_aw[0] + rgb_aw[1] + 0.05 * rgb_aw[2])
+        self.rgb_cw = alg.multiply_x3(self.rgb_w, self.d_rgb, dims=alg.D1)
+        rgb_aw = adapt(self.rgb_cw, self.fl)
+        self.a_w = (2 * rgb_aw[0] + rgb_aw[1] + 0.05 * rgb_aw[2])
 
 
 def cam_to_xyz(
@@ -101,13 +115,15 @@ def cam_to_xyz(
     s: float | None = None,
     Q: float | None = None,
     M: float | None = None,
+    Jhk: float | None = None,
+    Qhk: float | None = None,
     H: float | None = None,
     env: Environment | None = None
 ) -> Vector:
     """
-    From CAM02 to XYZ.
+    From CAM16 to XYZ.
 
-    Reverse calculation can actually be obtained from a small subset of the CAM02 components
+    Reverse calculation can actually be obtained from a small subset of the CAM16 components
     Really, only one suitable value is needed for each type of attribute: (lightness/brightness),
     (chroma/colorfulness/saturation), (hue/hue quadrature). If more than one for a given
     category is given, we will fail as we have no idea which is the right one to use. Also,
@@ -115,8 +131,8 @@ def cam_to_xyz(
     """
 
     # These check ensure one, and only one attribute for a given category is provided.
-    if not ((J is not None) ^ (Q is not None)):
-        raise ValueError("Conversion requires one and only one: 'J' or 'Q'")
+    if not ((J is not None) ^ (Q is not None) ^ (Jhk is not None) ^ (Qhk is not None)):
+        raise ValueError("Conversion requires one and only one: 'J', 'Q', 'Jhk', or 'Qhk'")
 
     if not ((C is not None) ^ (M is not None) ^ (s is not None)):
         raise ValueError("Conversion requires one and only one: 'C', 'M' or 's'")
@@ -130,7 +146,7 @@ def cam_to_xyz(
         raise ValueError("No viewing conditions/environment provided")
 
     # Black
-    if J == 0.0 or Q == 0.0:
+    if J == 0.0 or Q == 0.0 or Jhk == 0.0 or Qhk == 0.0:
         return [0.0, 0.0, 0.0]
 
     # Break hue into Cartesian components
@@ -141,87 +157,85 @@ def cam_to_xyz(
         h_rad = math.radians(inv_hue_quadrature(H))
 
     # Calculate `J_root` from one of the lightness derived coordinates.
-    J_root = 0.0
-    if J is not None:
-        J_root = alg.nth_root(J, 2) * 0.1
-    elif Q is not None:
-        J_root = 0.25 * env.c * Q / ((env.a_w + 4) * env.fl_root)
+    if M is not None:
+        C = M * 35 / env.a_w
+    elif C is not None:
+        M = (C * env.a_w) / 35
 
-    # Calculate the `t` value from one of the chroma derived coordinates
-    alpha = 0.0
-    if C is not None:
-        alpha = C / J_root
-    elif M is not None:
-        alpha = (M / env.fl_root) / J_root
-    elif s is not None:
-        alpha = 0.0004 * (s ** 2) * (env.a_w + 4) / env.c
-    t = alg.spow(alpha * math.pow(1.64 - math.pow(0.29, env.n), -0.73), 10 / 9)
+    if Qhk is not None:
+        Jhk = (50 * env.c * Qhk) / env.a_w
+
+    if Jhk is not None:
+        if C is None:
+            raise ValueError('C or M is required to resolve Jhk and Qhk')
+        J = Jhk - hue_angle_dependency(h_rad) * alg.spow(C, 0.587)
+
+    if J is not None:
+        Q = (2 / env.c) * (J / 100) * env.a_w
+    elif Q is not None:
+        J = (50 * env.c * Q) / env.a_w
+
+    if s is not None:
+        M = Q * (s / 100)  # type: ignore[operator]
 
     # Eccentricity
     et = eccentricity(h_rad)
 
     # Achromatic response
-    A = env.a_w * alg.spow(J_root, 2 / env.c / env.z)
+    A = env.a_w * alg.nth_root(J / 100, env.c * env.z)  # type: ignore[operator]
 
-    # Calculate red-green and yellow-blue components from hue
+    # Calculate red-green and yellow-blue components
     cos_h = math.cos(h_rad)
     sin_h = math.sin(h_rad)
-    p1 = 5e4 / 13 * env.nc * env.ncb * et
-    p2 = A / env.nbb
-    r = 23 * (p2 + 0.305) * alg.zdiv(t, 23 * p1 + t * (11 * cos_h + 108 * sin_h))
+    p1 = 43 * env.nc * et
+    p2 = A
+    r = M / p1  # type: ignore[operator]
     a = r * cos_h
     b = r * sin_h
 
     # Calculate back from cone response to XYZ
     rgb_a = alg.multiply_x3(alg.matmul_x3(M1, [p2, a, b], dims=alg.D2_D1), 1 / 1403, dims=alg.D1_SC)
-    rgb_c = alg.matmul_x3(alg.matmul_x3(M02, HPE_TO_XYZ, dims=alg.D2), unadapt(rgb_a, env.fl), dims=alg.D2_D1)
-    return util.scale1(alg.matmul_x3(M02_INV, alg.multiply_x3(rgb_c, env.d_rgb_inv, dims=alg.D1), dims=alg.D2_D1))
+    rgb_c = unadapt(rgb_a, env.fl)
+    return util.scale1(alg.matmul_x3(MI6_INV, alg.multiply_x3(rgb_c, env.d_rgb_inv, dims=alg.D1), dims=alg.D2_D1))
 
 
 def xyz_to_cam(xyz: Vector, env: Environment, calc_hue_quadrature: bool = False) -> Vector:
-    """From XYZ to CAM02."""
+    """From XYZ to CAM16."""
 
     # Calculate cone response
     rgb_c = alg.multiply_x3(
+        alg.matmul_x3(M16, util.scale100(xyz), dims=alg.D2_D1),
         env.d_rgb,
-        alg.matmul_x3(M02, util.scale100(xyz), dims=alg.D2_D1),
         dims=alg.D1
     )
-    rgb_a = adapt(alg.matmul_x3(alg.matmul_x3(XYZ_TO_HPE, M02_INV, dims=alg.D2), rgb_c, dims=alg.D2_D1), env.fl)
+    rgb_a = adapt(rgb_c, env.fl)
 
     # Calculate red-green and yellow components and resultant hue
     p2 = 2 * rgb_a[0] + rgb_a[1] + 0.05 * rgb_a[2]
     a = rgb_a[0] + (-12 * rgb_a[1] + rgb_a[2]) / 11
     b = (rgb_a[0] + rgb_a[1] - 2 * rgb_a[2]) / 9
-    u = rgb_a[0] + rgb_a[1] + 1.05 * rgb_a[2]
     h_rad = math.atan2(b, a) % math.tau
 
     # Eccentricity
     et = eccentricity(h_rad)
 
-    # Calculate `t` so we can calculate `alpha`
-    p1 = 5e4 / 13 * env.nc * env.ncb * et
-    t = alg.zdiv(p1 * math.sqrt(a ** 2 + b ** 2), u + 0.305)
-    alpha = alg.spow(t, 0.9) * math.pow(1.64 - math.pow(0.29, env.n), 0.73)
-
     # Achromatic response
-    A = env.nbb * p2
+    A = p2
 
     # Lightness
     J = 100 * alg.spow(A / env.a_w, env.c * env.z)
-    J_root = alg.nth_root(J / 100, 2)
 
     # Brightness
-    Q = (4 / env.c * J_root * (env.a_w + 4) * env.fl_root)
-
-    # Chroma
-    C = alpha * J_root
+    Q = (2 / env.c) * (J / 100) * env.a_w
 
     # Colorfulness
-    M = C * env.fl_root
+    M = 43 * env.nc * et * math.hypot(a, b)
+
+    # Chroma
+    C = 35 * (M / env.a_w)
 
     # Saturation
-    s = 50 * alg.nth_root(env.c * alpha / (env.a_w + 4), 2)
+    s = 100 * alg.zdiv(M, Q)
 
     # Hue
     h = util.constrain_hue(math.degrees(h_rad))
@@ -229,36 +243,44 @@ def xyz_to_cam(xyz: Vector, env: Environment, calc_hue_quadrature: bool = False)
     # Hue quadrature
     H = hue_quadrature(h) if calc_hue_quadrature else alg.NaN
 
-    return [J, C, h, s, Q, M, H]
+    # Lightness: Helmholtz-Kohlrausch effect
+    Jhk = J + hue_angle_dependency(h_rad) * alg.spow(C, 0.587)
+
+    # Brightness: Helmholtz-Kohlrausch effect
+    Qhk = (2 / env.c) * (Jhk / 100) * env.a_w
+
+    return [J, C, h, s, Q, M, Jhk, Qhk, H]
 
 
-def xyz_to_cam_jmh(xyz: Vector, env: Environment) -> Vector:
-    """XYZ to CAM02 JMh."""
+def xyz_to_cam_jmh(xyz: Vector, env: Environment, hk: bool = False) -> Vector:
+    """XYZ to CAM16 JMh with corrections."""
 
-    cam = xyz_to_cam(xyz, env)
-    J, M, h = cam[0], cam[5], cam[2]
+    cam16 = xyz_to_cam(xyz, env)
+    J, M, h = (cam16[6] if hk else cam16[0]), cam16[5], cam16[2]
     return [J, M, h]
 
 
-def cam_jmh_to_xyz(jmh: Vector, env: Environment) -> Vector:
-    """CAM02 JMh to XYZ."""
+def cam_jmh_to_xyz(jmh: Vector, env: Environment, hk: bool = False) -> Vector:
+    """CAM16 JMh with corrections to XYZ."""
 
     J, M, h = jmh
-    return cam_to_xyz(J=J, M=M, h=h, env=env)
+    return cam_to_xyz(Jhk=J, M=M, h=h, env=env) if hk else cam_to_xyz(J=J, M=M, h=h, env=env)
 
 
-class CAM02JMh(LChish, Space):
-    """CAM02 class (JMh)."""
+class HellwigJMh(LCh):
+    """CAM16 class (JMh) with corrections."""
 
     BASE = "xyz-d65"
-    NAME = "cam02-jmh"
-    SERIALIZE = ("--cam02-jmh",)
+    NAME = "hellwig-jmh"
+    SERIALIZE = ("--hellwig-jmh",)
     CHANNEL_ALIASES = {
         "lightness": "j",
         "colorfulness": 'm',
         "hue": 'h'
     }
     WHITE = WHITES['2deg']['D65']
+    HK = False
+
     # Assuming sRGB which has a lux of 64: `((E * R) / PI) / 5` where `R = 1`.
     ENV = Environment(
         # Our white point.
@@ -275,7 +297,7 @@ class CAM02JMh(LChish, Space):
     )
     CHANNELS = (
         Channel("j", 0.0, 100.0),
-        Channel("m", 0, 120.0),
+        Channel("m", 0, 70.0),
         Channel("h", 0.0, 360.0, flags=FLG_ANGLE)
     )
 
@@ -291,14 +313,43 @@ class CAM02JMh(LChish, Space):
         """Check if color is achromatic."""
 
         # Account for both positive and negative chroma
-        return coords[0] == 0 or abs(coords[1]) < ACHROMATIC_THRESHOLD
+        return coords[0] == 0 or abs(coords[1]) < self.achromatic_threshold
 
     def to_base(self, coords: Vector) -> Vector:
-        """From CAM02 JMh to XYZ."""
+        """From CAM16 JMh to XYZ."""
 
-        return cam_jmh_to_xyz(coords, self.ENV)
+        return cam_jmh_to_xyz(coords, self.ENV, hk=self.HK)
 
     def from_base(self, coords: Vector) -> Vector:
-        """From XYZ to CAM02 JMh."""
+        """From XYZ to CAM16 JMh."""
 
-        return xyz_to_cam_jmh(coords, self.ENV)
+        return xyz_to_cam_jmh(coords, self.ENV, hk=self.HK)
+
+
+class HellwigHKJMh(HellwigJMh):
+    """CAM16 class (JMh) with corrections and accounting for the Helmholtz-Kohlrausch effect."""
+
+    NAME = "hellwig-hk-jmh"
+    SERIALIZE = ("--hellwig-hk-jmh",)
+    WHITE = WHITES['2deg']['D65']
+    HK = True
+
+    # Assuming sRGB which has a lux of 64: `((E * R) / PI) / 5` where `R = 1`.
+    ENV = Environment(
+        # Our white point.
+        white=WHITE,
+        # Assuming sRGB which has a lux of 64: `((E * R) / PI)` where `R = 1`.
+        # Divided by 5 (or multiplied by 20%) assuming gray world.
+        adapting_luminance=64 / math.pi * 0.2,
+        # Gray world assumption, 20% of reference white's `Yw = 100`.
+        background_luminance=20,
+        # Average surround
+        surround='average',
+        # Do not discount illuminant
+        discounting=False
+    )
+    CHANNELS = (
+        Channel("j", 0.0, 101.56018891418564),
+        Channel("m", 0, 70.0),
+        Channel("h", 0.0, 360.0, flags=FLG_ANGLE)
+    )
