+++ docs/src/markdown/colors/zcam.md
@@ -81,7 +81,7 @@ When doing this, the space should be derived from the default
 
 ```py play
 from coloraide import Color as Base
-from coloraide.spaces.zcam_jmh import ZCAMJMh, Environment
+from coloraide.spaces.zcam import ZCAMJMh, Environment
 from coloraide.cat import WHITES
 from coloraide import util
 import math
@@ -138,7 +138,7 @@ Color("zcam-jmh", [71.271, 32.313, 75.038], 1).to_string()
 
 ```py
 from coloraide import Color as Base
-from coloraide.spaces.zcam_jmh import ZCAMJMh
+from coloraide.spaces.zcam import ZCAMJMh
 
 class Color(Base): ...
