+++ docs/src/mkdocs.yml
@@ -114,9 +114,9 @@ nav:
         - LCh(uv): colors/lchuv.md
         - DIN99 LCh: colors/lch99o.md
         - JzCzhz: colors/jzczhz.md
-        - CAM16 JMh: colors/cam16_jmh.md
+        - CAM16 JMh: colors/cam16.md
         - HCT: colors/hct.md
-        - ZCAM JMh: colors/zcam_jmh.md
+        - ZCAM JMh: colors/zcam.md
 
       - ACES Spaces:
         - ACES 2065-1: colors/aces2065_1.md
