+++ tools/calc_hct_j_poly.py
@@ -6,7 +6,7 @@ import numpy as np
 sys.path.insert(0, os.getcwd())
 
 from coloraide.everything import ColorAll as Color  # noqa: E402
-from coloraide.spaces import cam16_jmh  # noqa: E402
+from coloraide.spaces import cam16  # noqa: E402
 from coloraide.spaces import hct  # noqa: E402
 
 env = hct.HCT.ENV
@@ -16,7 +16,7 @@ t = []
 
 for r in range(200000):
     xyz = Color('srgb', [r / 100000] * 3).convert('xyz-d65')
-    j.append(cam16_jmh.xyz_to_cam(xyz.coords(), env)[0])
+    j.append(cam16.xyz_to_cam(xyz.coords(), env)[0])
     t.append(hct.y_to_lstar(xyz[1]))
 
 print('==== Positive Lightness ====')
@@ -27,7 +27,7 @@ t = []
 
 for r in range(200000):
     xyz = Color('srgb', [-r / 100000] * 3).convert('xyz-d65')
-    j.append(cam16_jmh.xyz_to_cam(xyz.coords(), env)[0])
+    j.append(cam16.xyz_to_cam(xyz.coords(), env)[0])
     t.append(hct.y_to_lstar(xyz[1]))
 
 print('==== Negative Lightness ====')
